# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bitcoin AI Analyst Backend",
    "version": "2026-09-research-evidence-v2"
  },
  "paths": {
...
    "/api/discovery/run": {
      "post": {
        "tags": [
          "RuleDiscovery"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "4h"
            }
          },
          {
            "name": "lookbackBars",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 3000
            }
          },
          {
            "name": "futureBars",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 3
            }
          },
          {
            "name": "minWinRate",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 0.5
            }
          },
          {
            "name": "minSamples",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          },
          {
            "name": "minAvgReturnPct",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 0
            }
          },
          {
            "name": "candidateBudget",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 128
            }
          },
          {
            "name": "selectionFraction",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 0.7
            }
          },
          {
            "name": "labelDeadZonePct",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 0.3
            }
          },
          {
            "name": "roundTripCostBps",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double",
              "default": 30
            }
          },
          {
            "name": "saveToDb",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/discovery/rules": {
      "get": {
        "tags": [
          "RuleDiscovery"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/discovery/index-volume": {
      "post": {
        "tags": [
          "RuleDiscovery"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "4h"
            }
          },
          {
            "name": "lookbackBars",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/discovery/volume-stats": {
      "get": {
        "tags": [
          "RuleDiscovery"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "4h"
            }
          },
          {
            "name": "take",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/discovery/evaluate": {
      "post": {
        "tags": [
          "RuleDiscovery"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "4h"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}